requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
future==0.16.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-telegram-bot==9.0.0
SQLAlchemy==1.2.4
Werkzeug==0.14.1
